Add confirmation popup & reservations search table view

+++ src/app/modules/admin/admin-reservation-management/component/admin-reservation-management/reservation-table/reservation-col/reservation-col.component.ts
@@ -1,8 +1,10 @@
-import {Component, Input, OnInit, Output} from '@angular/core';
+import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
 import {ReservationModel} from '../../../../../../../objects/models/ReservationModel';
 import {ReservationStatus} from '../../../../../../../objects/models/ReservationStatus';
 import {AdminReservationManagementApiService} from '../../../../services/admin-reservation-management-api.service';
-import {EventEmitter} from '@angular/core';
+import {isNotNullOrUndefined} from 'codelyzer/util/isNotNullOrUndefined';
+import {MatDialog} from '@angular/material/dialog';
+import {ConfirmationPopup} from '../../../../../../../utils/confirmation-popup/confirmation-popup';
 
 @Component({
   selector: 'app-reservation-col',
@@ -18,15 +20,25 @@ export class ReservationColComponent implements OnInit {
   @Output('onDetailsClick') onDetailsClickEventEmitter: EventEmitter<ReservationModel> = new EventEmitter();
   reservationStatuses = ReservationStatus;
 
-  constructor(private _reservationApiService: AdminReservationManagementApiService) {
+  constructor(private _reservationApiService: AdminReservationManagementApiService,
+              private _dialog: MatDialog) {
   }
 
   ngOnInit(): void {
   }
 
   removeReservation(value) {
-    this._reservationApiService.deleteReservation(value.id).subscribe(() => {
-      this.onDeleteItemEventEmitter.emit(value.id);
+    const dialogRef = this._dialog.open(ConfirmationPopup, {
+      width: '400px',
+      data: "Czy na pewno chcesz usunąć termin wizyty?"
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (isNotNullOrUndefined(result)) {
+        this._reservationApiService.deleteReservation(value.id).subscribe(() => {
+          this.onDeleteItemEventEmitter.emit(value.id);
+        });
+      }
     });
   }
